+++ src/Route.php
@@ -4,10 +4,29 @@ namespace Codemonster\Router;
 
 class Route
 {
-    public function __construct(
-        public array $methods,
-        public string $path,
-        public mixed $handler
-    ) {
+    public array $methods;
+    public string $path;
+    public mixed $handler;
+    protected array $middleware = [];
+
+    public function __construct(array|string $methods, string $path, mixed $handler)
+    {
+        $this->methods = (array)$methods;
+        $this->path = $path;
+        $this->handler = $handler;
+    }
+
+    public function middleware(string|array $middleware): static
+    {
+        foreach ((array)$middleware as $m) {
+            $this->middleware[] = $m;
+        }
+
+        return $this;
+    }
+
+    public function getMiddleware(): array
+    {
+        return $this->middleware;
     }
 }